# artifacts/api-server/src/routes/programs.ts
import { Router } from "express";
import { z } from "zod";
import { prisma } from "../lib/prisma.js";
import { authenticate } from "../middlewares/auth.js";
import { authorizePermission } from "../middlewares/rbac.js";
import { requireTenant, assertTenantMatch } from "../middlewares/tenantScope.js";
import { validateBody } from "../middlewares/validate.js";
import { Errors, paginate, paginationMeta } from "../types/index.js";
import { createAuditLog } from "../lib/audit.js";
import { getRoleScope } from "../middlewares/roleScope.js";

const router = Router();
router.use(authenticate, requireTenant);

const ProgramSchema = z.object({
  tenantId: z.string().uuid().optional(),
  areaId: z.string().uuid().optional().nullable(),
  clinicId: z.string().uuid().optional().nullable(),
  name: z.string().min(1).max(100),
  description: z.string().max(500).optional().nullable(),
  activationCode: z.string().max(50).optional().nullable(),
  channelId: z.number().int().optional().nullable(),
  logoUrl: z.string().url().optional().nullable(),
  priority: z.number().int().min(0).optional(),
  tags: z.array(z.string()).optional(),
  isActive: z.boolean().optional(),
});

router.get("/", authorizePermission("programs", "read"), async (req, res, next) => {
  try {
    const { skip, take, page, limit } = paginate(req.query);
    const q = req.query["q"] as string | undefined;
    const reqTenantId = req.query["tenantId"] as string | undefined;
    const areaId = req.query["areaId"] as string | undefined;
    const clinicId = req.query["clinicId"] as string | undefined;
    const roleScope = await getRoleScope(req, "program");
    const where = {
      deletedAt: null,
      ...(req.tenantId ? { tenantId: req.tenantId } : reqTenantId ? { tenantId: reqTenantId } : {}),
      ...roleScope,
      ...(areaId ? { areaId } : {}),
      ...(clinicId ? { clinicId } : {}),
      ...(q ? { name: { contains: q, mode: "insensitive" as const } } : {}),
    };
    const [total, programs] = await Promise.all([
      prisma.program.count({ where }),
      prisma.program.findMany({
        where, skip, take, orderBy: [{ priority: "asc" }, { name: "asc" }],
        include: { _count: { select: { patients: true } } },
      }),
    ]);
    res.json({ data: programs, meta: paginationMeta(total, page, limit) });
  } catch (err) { next(err); }
});

router.get("/:id", authorizePermission("programs", "read"), async (req, res, next) => {
  try {
    const program = await prisma.program.findFirst({
      where: { id: req.params["id"] as string, tenantId: req.tenantId!, deletedAt: null },
      include: { _count: { select: { patients: true } } },
    });
    if (!program) throw Errors.notFound("Program");
    res.json(program);
  } catch (err) { next(err); }
});

router.post("/", authorizePermission("programs", "write"), validateBody(ProgramSchema), async (req, res, next) => {
  try {
    const data = req.body as z.infer<typeof ProgramSchema>;
    const targetTenantId = data.tenantId || req.tenantId;
    if (!targetTenantId) throw Errors.badRequest("Tenant ID is required");

    const activationCode = data.activationCode ?? `${data.name.toUpperCase().replace(/\s+/g, "_").slice(0, 20)}_${Date.now()}`;

    // Validate hierarchy and tenant ownership
    if (data.areaId) {
      const area = await prisma.area.findFirst({ where: { id: data.areaId, tenantId: targetTenantId, deletedAt: null } });
      if (!area) throw Errors.validation("Area does not belong to this tenant");
    }
    if (data.clinicId) {
      const clinicWhere: any = { id: data.clinicId, tenantId: targetTenantId, deletedAt: null };
      if (data.areaId) clinicWhere.areaId = data.areaId;
      const clinic = await prisma.clinic.findFirst({ where: clinicWhere });
      if (!clinic) throw Errors.validation("Clinic does not belong to this tenant/area");
    }

    const program = await prisma.program.create({ data: { ...data, activationCode, tenantId: targetTenantId, areaId: data.areaId, clinicId: data.clinicId } });
    await createAuditLog({ req, entityType: "Program", entityId: program.id, action: "CREATE", after: data });
    res.status(201).json(program);
  } catch (err) { next(err); }
});

router.patch("/:id", authorizePermission("programs", "write"), validateBody(ProgramSchema.partial()), async (req, res, next) => {
  try {
    const program = await prisma.program.findFirst({ where: { id: req.params["id"] as string, deletedAt: null } });
    if (!program) throw Errors.notFound("Program");
    assertTenantMatch(req, program.tenantId);

    const targetTenantId = program.tenantId;
    if (req.body.areaId) {
      const area = await prisma.area.findFirst({ where: { id: req.body.areaId, tenantId: targetTenantId, deletedAt: null } });
      if (!area) throw Errors.validation("Area does not belong to this tenant");
    }
    if (req.body.clinicId) {
      const areaId = req.body.areaId ?? program.areaId;
      const clinicWhere: any = { id: req.body.clinicId, tenantId: targetTenantId, deletedAt: null };
      if (areaId) clinicWhere.areaId = areaId;
      const clinic = await prisma.clinic.findFirst({ where: clinicWhere });
      if (!clinic) throw Errors.validation("Clinic does not belong to this tenant/area");
    }

    const updated = await prisma.program.update({ where: { id: program.id }, data: req.body });
    await createAuditLog({ req, entityType: "Program", entityId: program.id, action: "UPDATE", before: program, after: req.body });
    res.json(updated);
  } catch (err) { next(err); }
});

router.delete("/:id", authorizePermission("programs", "write"), async (req, res, next) => {
  try {
    const program = await prisma.program.findFirst({ where: { id: req.params["id"] as string, deletedAt: null } });
    if (!program) throw Errors.notFound("Program");
    assertTenantMatch(req, program.tenantId);
    await prisma.program.update({ where: { id: program.id }, data: { deletedAt: new Date() } });
    await createAuditLog({ req, entityType: "Program", entityId: program.id, action: "DELETE" });
    res.status(204).send();
  } catch (err) { next(err); }
});

export default router;