# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareNexus API — Connected Care. Better Outcomes.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramList"
    post:
      operationId: createProgram
      tags: [programs]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    patch:
      operationId: updateProgram
      tags: [programs]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [programs]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Patients ──────────────────────────────────────────────────────────────
...
  /program-enrollments:
    get:
      operationId: listProgramEnrollments
      tags: [programEnrollments]
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: programId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of program enrollments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramEnrollmentList"
    post:
      operationId: createProgramEnrollment
      tags: [programEnrollments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramEnrollmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramEnrollment"
  /program-enrollments/{id}:
    patch:
      operationId: updateProgramEnrollment
      tags: [programEnrollments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramEnrollmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramEnrollment"
  /program-enrollments/{id}/complete:
    post:
      operationId: completeProgramEnrollment
      tags: [programEnrollments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramEnrollment"
  /program-enrollments/{id}/cancel:
    post:
      operationId: cancelProgramEnrollment
      tags: [programEnrollments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramEnrollment"

  # ── Appointments ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    Meta:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Program:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    ProgramInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    ProgramUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ProgramList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Program"
        meta:
          $ref: "#/components/schemas/Meta"

    # ── Patient ─────────────────────────────────────────────────────────────
...
    ProgramEnrollment:
      type: object
      required: [id, tenantId, patientId, programId, status, enrolledAt, createdAt, updatedAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        patientId:
          type: string
        programId:
          type: string
        enrolledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ACTIVE, COMPLETED, CANCELLED]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        program:
          $ref: "#/components/schemas/Program"
    ProgramEnrollmentList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProgramEnrollment"
        meta:
          $ref: "#/components/schemas/Meta"
    ProgramEnrollmentInput:
      type: object
      required: [patientId, programId]
      properties:
        patientId:
          type: string
        programId:
          type: string
        notes:
          type: string
    ProgramEnrollmentUpdate:
      type: object
      properties:
        notes:
          type: string
...